LayoutTests/editing/spelling/spelling-should-select-single-word-expected.txt

Spell check selects the misspelled word on context click. To test manually, type 'wellcome home.' and context-click on 'wellcome'. The test succeeds when 'wellcome' is selected after the context click.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is "wellcome"
PASS contextMenuElements[contextMenuElements.length - 1] is "welcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is ""
PASS contextMenuElements[contextMenuElements.length - 1] is "<separator>"
PASS successfullyParsed is true

TEST COMPLETE
wellcome home.
